Size file size into 2 fields
I don't want to drop the file size maximum below 40 bits (1 terrabyte)
but I also don't want to allow the compiler to add 3 more bytes.
It wants to round 5 bytes to 8 bytes.
So split file size into 2 fields, the first being the number of times the file exceeds 4Gb.
For 99.9% of the files this will always be zero.
The "modulus" part *is* the file size for all of these.