Next iteration
There have been a lot of changes.
What I don't like is now instead of no temporary files, we have two:
The original plain archive and now the old blocks 2 .. 5

But the tradeoff is now everything is compressed except the first 32 bytes.
For archives with 10,000+ files, this will make a huge difference in file size.
Now to test it ....