The problem is 24-bit numbers
Even though Ada considers this 24 bits, it always takes 32 bits storage.
Stop being stingy and use one of the 4-byte unused blocks.
Using the same logic of not trying to pack so tightly, set the flat
archives to use 8 bytes instead of 5, which allows us to keep it in one
variable again (hopefully).