drop %F expansion
It doesn't really make sense.
If this is a [file] or [dir] action, the equivalent to %F will be the same as %1, so use it instead.

I'm not going to have the extractor keep track of the "last filename extracted.
Apparently previous tools used "cd" which affects the prefix (%D would be variable).
Currently no freebsd keyword uses %F so I'm treating it as obsolete.