Remove use of extensions.bool
I'm not sure if it's a compiler bug or what, but somehow all the uses
of bool were considered fixed by the compiler, so switch to unsigned_char instead.
It actually makes the code a little cleaner.

Also, ucl_object_iterate is not a real symbol so change it to it's real
name: ucl_object_iterate_with_error  to fix linking.