We're not done with Lua yet
I just figured out that the arguments are supposed to be stored along with the Lua code.
That requires a change in metadata structure.

This code should implement it.  The bourne scripts do NOT store the arguments.
For shell scripts, the arguments are replaced directly in the code at pack time.